Update to pubid-core 0.5
