+++ pubid-ieee.gemspec
@@ -31,5 +31,5 @@ Gem::Specification.new do |spec|
   spec.add_dependency "thor"
   spec.add_dependency "lightly"
   spec.add_dependency "parslet"
-  spec.add_dependency "pubid-iso", "~> 0.4.1"
+  spec.add_dependency "pubid-iso", "~> 0.5.2"
 end
